requirements.txt
tqdm==4.47.0
selenium==3.141.0
Pillow==8.3.1
